更改了i2d.py

+++ ReasoningTrack/Reinforcement Learning/scripts/model_merger.py
@@ -174,7 +174,8 @@ if __name__ == "__main__":
         raise NotImplementedError(f"Unknown architecture {architectures}.")
 
     with torch.device("meta"):
-        model: PreTrainedModel = AutoClass.from_config(config, torch_dtype=torch.bfloat16)
+        # model: PreTrainedModel = AutoClass.from_config(config, torch_dtype=torch.bfloat16)
+        model: PreTrainedModel = AutoClass.from_config(config)
 
     assert isinstance(model, PreTrainedModel)
     model.to_empty(device="cpu")
